requirements.txt
azure-functions
azure-search-documents==11.6.0b4
azure-identity
azure-storage-blob
openai
tenacity
langchain-text-splitters
azure-monitor-opentelemetry
PyPDF2
qdrant-client==1.12.0
grpcio==1.62.1
